refactor: directory arrangement, env
