+++ .env
@@ -1,0 +1,2 @@
+# development
+REACT_APP_DEV_API_URL="http://localhost:3001"
