+++ src/components/Event.js
@@ -18,14 +18,13 @@ const Event = () => {
         handleGetEvent();
     }, [])
 
-    const imageURL = `http://localhost:3001`
     return (
         <div>
             {events.map((event) => (
                 <div key={event.id}>
                     <p>{event.title}</p>
                     <p>{event.date}</p>
-                    <img src={imageURL + event.image.url} alt="event_image"/>
+                    <img src={process.env.REACT_APP_DEV_API_URL + event.image.url} alt="event_image"/>
                 </div>
             ))}
         </div>
